requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
qdrant-client
sentence-transformers
pymupdf
requests
python-multipart
tf-keras
google-generativeai
